+++ src/main/java/mod/icarus/crimsonrevelations/block/BlockManaPod.java
@@ -1,7 +1,9 @@
 package mod.icarus.crimsonrevelations.block;
 
+import com.invadermonky.magicultureintegrations.api.block.IHarvestableCrop;
 import mod.icarus.crimsonrevelations.NewCrimsonRevelations;
 import mod.icarus.crimsonrevelations.config.ConfigHandlerNCR;
+import mod.icarus.crimsonrevelations.registry.ModBlocksNCR;
 import mod.icarus.crimsonrevelations.registry.ModItemsNCR;
 import mod.icarus.crimsonrevelations.item.misc.ItemManaBean;
 import mod.icarus.crimsonrevelations.tile.TileManaPod;
@@ -12,6 +14,7 @@ import net.minecraft.block.material.Material;
 import net.minecraft.block.properties.PropertyInteger;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
@@ -24,36 +27,37 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 import net.minecraftforge.common.BiomeDictionary;
+import net.minecraftforge.common.util.Constants;
+import net.minecraftforge.fml.common.Optional;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import thaumcraft.api.aspects.Aspect;
 import thaumcraft.api.aspects.AspectList;
 import thaumcraft.api.blocks.BlocksTC;
 import thaumcraft.api.internal.WorldCoordinates;
+import vazkii.botania.api.item.IHornHarvestable;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.Objects;
 import java.util.Random;
 
 @SuppressWarnings("deprecation")
-public class BlockManaPod extends Block implements IGrowable {
+@Optional.InterfaceList({
+        @Optional.Interface(modid = "botania", iface = "vazkii.botania.api.item.IHornHarvestable"),
+        @Optional.Interface(modid = "magicultureintegrations", iface = "com.invadermonky.magicultureintegrations.api.block.IHarvestableCrop")
+})
+public class BlockManaPod extends Block implements IGrowable, IHarvestableCrop, IHornHarvestable {
     public static final PropertyInteger AGE = PropertyInteger.create("age", 0, 8);
-    public static float W1 = 0.0625F;
-    public static float W2 = 0.125F;
-    public static float W3 = 0.1875F;
-    public static float W4 = 0.25F;
-    public static float W5 = 0.3125F;
-    public static float W6 = 0.375F;
-    public static float W7 = 0.4375F;
-    public static float W8 = 0.5F;
-    public static float W9 = 0.5625F;
-    public static float W10 = 0.625F;
-    public static float W11 = 0.6875F;
-    public static float W12 = 0.75F;
-    public static float W13 = 0.8125F;
-    public static float W14 = 0.875F;
-    public static float W15 = 0.9375F;
+    public static float W1 = 0.125F;
+    public static float W2 = 0.1875F;
+    public static float W3 = 0.25F;
+    public static float W4 = 0.3125F;
+    public static float W5 = 0.375F;
+    public static float W6 = 0.5F;
+    public static float W7 = 0.625F;
+    public static float W8 = 0.75F;
     static HashMap<WorldCoordinates, Aspect> st = new HashMap<>();
 
     public BlockManaPod(String unlocName) {
@@ -125,24 +129,24 @@ public class BlockManaPod extends Block implements IGrowable {
 
         switch (l) {
             case 0:
-                return new AxisAlignedBB(0.25F, W12, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 1:
-                return new AxisAlignedBB(0.25F, W10, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 2:
                 return new AxisAlignedBB(0.25F, W8, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 3:
+            case 1:
+                return new AxisAlignedBB(0.25F, W7, 0.25F, 0.75F, 1.0F, 0.75F);
+            case 2:
                 return new AxisAlignedBB(0.25F, W6, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 4:
+            case 3:
                 return new AxisAlignedBB(0.25F, W5, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 5:
+            case 4:
                 return new AxisAlignedBB(0.25F, W4, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 6:
+            case 5:
                 return new AxisAlignedBB(0.25F, W3, 0.25F, 0.75F, 1.0F, 0.75F);
-            case 7:
+            case 6:
                 return new AxisAlignedBB(0.25F, W2, 0.25F, 0.75F, 1.0F, 0.75F);
+            case 7:
+                return new AxisAlignedBB(0.25F, W1, 0.25F, 0.75F, 1.0F, 0.75F);
         }
 
-        return new AxisAlignedBB(0.25F, W2, 0.25F, 0.75F, 1.0F, 0.75F);
+        return new AxisAlignedBB(0.25F, W1, 0.25F, 0.75F, 1.0F, 0.75F);
     }
 
     @SideOnly(Side.CLIENT)
@@ -286,5 +290,61 @@ public class BlockManaPod extends Block implements IGrowable {
     @Override
     public void grow(World world, @Nonnull Random random, @Nonnull BlockPos pos, IBlockState state) {
         world.setBlockState(pos, state.withProperty(AGE, state.getValue(AGE) + 1), 8);
+    }
+
+    //##########################################################
+    // Botania Integration
+    @Override
+    public boolean canHornHarvest(World world, BlockPos blockPos, ItemStack itemStack, EnumHornType enumHornType) {
+        return enumHornType == EnumHornType.WILD;
+    }
+
+    @Override
+    public boolean hasSpecialHornHarvest(World world, BlockPos blockPos, ItemStack itemStack, EnumHornType enumHornType) {
+        return enumHornType == EnumHornType.WILD;
+    }
+
+    @Override
+    public void harvestByHorn(World world, BlockPos blockPos, ItemStack itemStack, EnumHornType enumHornType) {
+        IBlockState state = world.getBlockState(blockPos);
+        if (state.getBlock() == ModBlocksNCR.MANA_POD && enumHornType == EnumHornType.WILD) {
+            if (state.getValue(AGE) >= 8) {
+                world.setBlockState(blockPos, state.getBlock().getDefaultState());
+                world.playEvent(Constants.WorldEvents.BREAK_BLOCK_EFFECTS, blockPos, Block.getStateId(state));
+                NonNullList<ItemStack> drops = NonNullList.create();
+                state.getBlock().getDrops(drops, world, blockPos, state, 0);
+                drops.forEach(stack -> Block.spawnAsEntity(world, blockPos, stack));
+            }
+        }
+    }
+
+    //##########################################################
+    // Magiculture Integrations Integration
+
+    @Override
+    public BlockPos getHarvestPosition(World world, BlockPos cropPos) {
+        return cropPos;
+    }
+
+    @Nonnull
+    @Override
+    public HarvestResult getHarvestResult(World world, BlockPos pos) {
+        IBlockState state = world.getBlockState(pos);
+        if (state.getBlock() instanceof IHarvestableCrop) {
+            return this.getAge(state) >= 8 ? HarvestResult.HARVEST : HarvestResult.CLAIM;
+        }
+        return HarvestResult.PASS;
+    }
+
+    @Nonnull
+    @Override
+    public NonNullList<ItemStack> harvestCrop(@Nullable EntityPlayer entityPlayer, World world, BlockPos pos, boolean silkTouch, int fortune) {
+        IBlockState state = world.getBlockState(pos);
+        NonNullList<ItemStack> drops = NonNullList.create();
+        if (state.getBlock() instanceof IHarvestableCrop && ((IHarvestableCrop) state.getBlock()).getHarvestResult(world, pos) == HarvestResult.HARVEST) {
+            state.getBlock().getDrops(drops, world, pos, state, fortune);
+            world.setBlockState(pos, state.getBlock().getDefaultState());
+        }
+        return drops;
     }
 }